requirements.txt
boto3==1.7.64
Django==2.0
django-allauth==0.36.0
django-cors-headers==2.4.0
django-filter==2.0.0
django-rest-auth==0.9.3
django-storages==1.6.6
djangorestframework==3.8.2
Markdown==2.6.11
python-dateutil==2.7.3
gunicorn==19.9.0
whitenoise==3.3.1
dj-database-url==0.5.0
python-decouple==3.1
psycopg2-binary==2.7.5
django_rest_swagger==2.2.0